requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pymupdf>=1.24
python-dotenv>=1.0
google-genai>=0.3
pydantic>=2.7